refactor(Timepicker): #635
