+++ components/datepicker/helpers.ts
@@ -63,14 +63,14 @@ export function createDate(date: string) {
     return new Date(date);
 }
 
-export function findValueIndex(values: StateValueItem[], value: StateValueItem) {
+export function findValueIndex(values: StateValueItem[], value: StateValueItem, type: OpUnitType) {
     return values.findIndex(item => {
         if (Array.isArray(item)) {
             // is multipe range values
             const [start, end] = value as [Dayjs, Dayjs];
-            return isEqual(item[0], start) && isEqual(item[1], end);
+            return isEqual(item[0], start, type) && isEqual(item[1], end, type);
         } else {
-            return isEqual(value as Dayjs, item);
+            return isEqual(value as Dayjs, item, type);
         }
     });
 }
