+++ components/timepicker/useValue.ts
@@ -1,12 +1,10 @@
 import {useInstance} from 'intact';
 import type {Timepicker} from './';
-import {useState} from '../../hooks/useState';
-import dayjs, {Dayjs} from 'dayjs';
 import type {useFormats} from './useFormats';
-import {isEqualArray, last} from '../utils';
-import {useValue as useValueBase} from '../datepicker/useValue';
+import {last} from '../utils';
+import {useValue as useValueBase, StateValueItem, DayjsValueItem} from '../datepicker/basepicker';
 import type {useDisabled} from '../datepicker/useDisabled';
-import {PanelTypes, PanelFlags, usePanel} from '../datepicker/usePanel';
+import type {usePanel} from '../datepicker/usePanel';
 import {NOW} from './constants';
 
 export function useValue(
@@ -15,9 +13,34 @@ export function useValue(
     panel: ReturnType<typeof usePanel>,
 ) {
     const instance = useInstance() as Timepicker;
-    const {value, ...rest} = useValueBase(formats, disabled, panel);
+    const {value, getDayjsValue, setValue, ...rest} = useValueBase(
+        formats,
+        disabled,
+        panel,
+        function(v: StateValueItem) {
+            return false;
+        },
+        function(v: DayjsValueItem) {
+            // pop the last value when input
+            // because it is a placeholder date
+            value.value.pop();
+            setValue(v, true);
+        },
+        function() {
+            return 'second';
+        } 
+    );
 
-    value.set([NOW]);
+    setValue(getPlaceholderDates(), false);
 
-    return {value, ...rest};
+    instance.watch('value', value => {
+        const dayjsValue = getDayjsValue(); 
+        setValue(last(dayjsValue) || getPlaceholderDates(), false);
+    });
+
+    function getPlaceholderDates(): StateValueItem {
+        return instance.get('range') ? [NOW, NOW] : NOW;
+    }
+
+    return {value, setValue, getDayjsValue,  ...rest};
 }
