+++ demos/components/datepicker/index.ts
@@ -1,0 +1,11 @@
+import {Component} from 'intact';
+import template from './index.vdt';
+
+export default class extends Component {
+    static template = template;
+    static defaults() {
+        return {
+            days: [],
+        };
+    }
+}
