+++ demos/components/select/index.vdt
@@ -1,7 +1,7 @@
 import {Select, Option} from 'kpc/components/select';
 
 <div>
-    <Select v-model="days" multiple>
+    <Select v-model="days" multiple filterable>
         <Option value="Monday">星期一</Option>
         <Option value="Tuesday">星期二</Option>
         <Option value="Wednesday">星期三</Option>
